requirements.txt
folium==0.20.0
haversine==2.9.0
inflection==0.5.1
pandas==3.0.1
Pillow==12.1.1
plotly==6.3.0
streamlit==1.51.0
streamlit_folium==0.26.2
